requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
discord.py==2.6.4
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
nats-py==2.13.1
propcache==0.4.1
python-dotenv==1.2.1
yarl==1.22.0
fastapi>=0.104.0
uvicorn>=0.24.0
pyyaml>=6.0
psutil>=5.9.0
requests>=2.32.0
tenacity>=8.2.0
